+++ packages/terra-dropdown-button/tests/wdio/dropdown-button-spec.js
@@ -6,7 +6,9 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
       browser.moveToObject('#root', 100, 1);
     });
 
-    Terra.it.validatesElement();
+    it('displays default Dropdown Button', () => {
+      Terra.validates.element('default');
+    });
 
     it('validates focus styling', () => {
       browser.keys('Tab');
@@ -18,12 +20,12 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
     it('opens and validates dropdown with click', () => {
       browser.click('[class*=dropdown-button]');
       Terra.validates.element('dropdown open', { selector: '[class*=dropdown-list]' });
-      Terra.validates.screenshot('dropdown button with dropdown open');
+      Terra.validates.element('dropdown button with dropdown open');
     });
 
     it('prints MetaData of option on click', () => {
       browser.click('#opt1');
-      Terra.validates.screenshot('MetaData of 1st option', { selector: '#root' });
+      Terra.validates.element('MetaData of 1st option', { selector: '#root' });
     });
   });
 
@@ -32,11 +34,13 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
       browser.url('/#/raw/tests/terra-dropdown-button/dropdown-button/disabled-dropdown-button');
     });
 
-    Terra.it.validatesElement();
+    it('displays disabled Dropdown Button', () => {
+      Terra.validates.element('default');
+    });
 
     it('tries to tab to the button', () => {
       browser.keys(['Tab']);
-      Terra.validates.screenshot('tab attempted');
+      Terra.validates.element('tab attempted');
     });
 
     it('tries to click the button', () => {
@@ -51,7 +55,9 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
       browser.moveToObject('#root', 100, 1);
     });
 
-    Terra.it.matchesScreenshot();
+    it('displays Dropdown Button with interactions', () => {
+      Terra.validates.element('default');
+    });
 
     it('opens the dropdown with click', () => {
       browser.click('[class*=dropdown-button]');
@@ -60,7 +66,7 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
 
     it('closes dropdown on escape without running a callback', () => {
       browser.keys(['Escape']);
-      Terra.validates.screenshot('escape');
+      Terra.validates.element('escape');
     });
 
     it('opens the dropdown with enter', () => {
@@ -85,69 +91,69 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
 
       it('runs callback on space', () => {
         browser.keys(['ArrowDown', 'Space']);
-        Terra.validates.screenshot('space');
+        Terra.validates.element('space');
       });
 
       it('keyboard navigates down and runs callback on space', () => {
         browser.keys(['ArrowDown', 'ArrowDown', 'Space']);
-        Terra.validates.screenshot('down arrow');
+        Terra.validates.element('down arrow');
       });
 
       it('runs callback on enter', () => {
         browser.keys(['ArrowDown', 'ArrowDown', 'ArrowDown', 'Enter']);
-        Terra.validates.screenshot('enter');
+        Terra.validates.element('enter');
       });
 
       it('runs callback on click', () => {
         browser.click('#red');
-        Terra.validates.screenshot('click');
+        Terra.validates.element('click');
       });
 
       it('will not keyboard navigates down past the last option', () => {
         browser.keys(['ArrowDown', 'ArrowDown', 'ArrowDown', 'ArrowDown', 'Space']);
-        Terra.validates.screenshot('no down out of bounds');
+        Terra.validates.element('no down out of bounds');
       });
 
       it('keyboard navigates up and runs callback on enter', () => {
         browser.keys(['ArrowDown', 'ArrowDown', 'ArrowUp', 'Enter']);
-        Terra.validates.screenshot('up arrow');
+        Terra.validates.element('up arrow');
       });
 
       it('will not keyboard navigates up past the first option', () => {
         browser.keys(['ArrowDown', 'ArrowDown', 'ArrowUp', 'ArrowUp', 'Enter']);
-        Terra.validates.screenshot('no up out of bounds');
+        Terra.validates.element('no up out of bounds');
       });
 
       it('jumps to the last entry', () => {
         browser.keys(['ArrowDown', 'End', 'Enter']);
-        Terra.validates.screenshot('end');
+        Terra.validates.element('end');
       });
 
       it('jumps to the first entry', () => {
         browser.keys(['ArrowDown', 'Home', 'Enter']);
-        Terra.validates.screenshot('home');
+        Terra.validates.element('home');
       });
 
       it('jumps when typing', () => {
         browser.keys(['ArrowDown', 'b', 'Enter']);
-        Terra.validates.screenshot('jumps when typing');
+        Terra.validates.element('jumps when typing');
       });
 
       it('closes on tab without running a callback', () => {
         browser.keys(['Tab']);
-        Terra.validates.screenshot('tab');
+        Terra.validates.element('tab');
       });
 
       it('closes the dropdown when clicking the caret with the dropdown open', () => {
         browser.click('[class*=dropdown-button]');
-        Terra.validates.screenshot('clicking caret closes dropdown');
+        Terra.validates.element('clicking caret closes dropdown');
       });
 
       it('closes the dropdown when clicking outside the dropdownbutton', () => {
         browser.moveToObject('#root', 200, 1);
         browser.buttonDown();
         browser.buttonUp();
-        Terra.validates.screenshot('clicking outside closes dropdown');
+        Terra.validates.element('clicking outside closes dropdown');
       });
     });
 
@@ -168,7 +174,9 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
       browser.moveToObject('#root', 1, 300);
     });
 
-    Terra.it.validatesElement('Both closed');
+    it('displays wide contents Dropdown Button both closed', () => {
+      Terra.validates.element('Both closed');
+    });
 
     it('opens the long label dropdown', () => {
       browser.click('#wide-label');
@@ -182,12 +190,10 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
     });
   });
 
-  describe('Block', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-dropdown-button/dropdown-button/block-dropdown-button');
-    });
+  it('displays block Dropdown Button', () => {
+    browser.url('/#/raw/tests/terra-dropdown-button/dropdown-button/block-dropdown-button');
 
-    Terra.it.validatesElement();
+    Terra.validates.element('block');
   });
 
   describe('Bounded', () => {
@@ -231,16 +237,14 @@ Terra.describeViewports('Dropdown Button', ['medium'], () => {
 
     it('runs callback in dropdown that focuses an element', () => {
       browser.click('#first');
-      Terra.validates.screenshot('first button clicked');
+      Terra.validates.element('first button clicked');
     });
   });
 
-  describe('Variants', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-dropdown-button/dropdown-button/variant-dropdown-buttons');
-      browser.moveToObject('#root', 1, 100);
-    });
+  it('displays variants of Dropdown Button', () => {
+    browser.url('/#/raw/tests/terra-dropdown-button/dropdown-button/variant-dropdown-buttons');
+    browser.moveToObject('#root', 1, 100);
 
-    Terra.it.validatesElement();
+    Terra.validates.element('variants');
   });
 });